feat(ee): move billing UI and API client to ee/features/billing
