+++ frontend/ui/src/features/settings/workspace/components/index.ts
@@ -4,5 +4,4 @@
 export { GeneralTab } from "./GeneralTab";
 export { MembersTab } from "./MembersTab";
 export { IntegrationsTab } from "./IntegrationsTab";
-export { BillingTab } from "./BillingTab";
 export { ModelProvidersTab } from "./ModelProvidersTab";
